requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.5
djangogrpcframework==0.2.1
djangorestframework==3.13.1
grpcio==1.46.3
grpcio-tools==1.46.3
protobuf==3.20.1
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
